backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pillow==11.0.0
openai==2.6.1
supabase==2.22.3
python-dotenv==1.0.0
pydantic==2.12.3
httpx==0.28.1
websockets==15.0.1
